Rewrite code so that it is easier to vectorize. With these changes we now beat stock bzip2 handsomely on the compression benchmarks that I looked at. e.g

```
Benchmark 1 (6 runs): target/release/examples/compress c 9 /home/folkertdev/rust/zlib-rs/silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           923ms ± 9.69ms     918ms …  943ms          0 ( 0%)        0%
  peak_rss           29.5MB ± 98.7KB    29.4MB … 29.6MB          0 ( 0%)        0%
  cpu_cycles         4.16G  ± 33.1M     4.14G  … 4.23G           0 ( 0%)        0%
  instructions       7.94G  ±  310      7.94G  … 7.94G           0 ( 0%)        0%
  cache_references    150M  ±  504K      150M  …  151M           0 ( 0%)        0%
  cache_misses       37.8M  ±  445K     37.1M  … 38.4M           0 ( 0%)        0%
  branch_misses      61.2M  ± 46.2K     61.2M  … 61.3M           0 ( 0%)        0%
Benchmark 2 (7 runs): target/release/examples/compress rs 9 /home/folkertdev/rust/zlib-rs/silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           818ms ± 3.00ms     814ms …  823ms          0 ( 0%)        ⚡- 11.5% ±  0.9%
  peak_rss           29.6MB ± 49.5KB    29.5MB … 29.6MB          1 (14%)          +  0.3% ±  0.3%
  cpu_cycles         3.67G  ± 10.8M     3.66G  … 3.69G           0 ( 0%)        ⚡- 11.8% ±  0.7%
  instructions       9.27G  ± 40.5K     9.27G  … 9.27G           0 ( 0%)        💩+ 16.7% ±  0.0%
  cache_references    151M  ±  556K      150M  …  152M           0 ( 0%)          +  0.5% ±  0.4%
  cache_misses       38.4M  ±  302K     38.0M  … 38.9M           0 ( 0%)          +  1.6% ±  1.2%
  branch_misses      52.3M  ± 21.3K     52.3M  … 52.4M           0 ( 0%)        ⚡- 14.5% ±  0.1%
```